--- a/chainer/functions/__init__.py
+++ b/chainer/functions/__init__.py
@@ -45,6 +45,7 @@ from chainer.functions.connection import convolution_2d  # NOQA
 from chainer.functions.connection import convolution_nd  # NOQA
 from chainer.functions.connection import deconvolution_2d  # NOQA
 from chainer.functions.connection import deconvolution_nd  # NOQA
+from chainer.functions.connection import depthwise_convolution_2d  # NOQA
 from chainer.functions.connection import dilated_convolution_2d  # NOQA
 from chainer.functions.connection import embed_id  # NOQA
 from chainer.functions.connection import linear  # NOQA
@@ -191,6 +192,7 @@ from chainer.functions.connection.convolution_2d import convolution_2d  # NOQA
 from chainer.functions.connection.convolution_nd import convolution_nd  # NOQA
 from chainer.functions.connection.deconvolution_2d import deconvolution_2d  # NOQA
 from chainer.functions.connection.deconvolution_nd import deconvolution_nd  # NOQA
+from chainer.functions.connection.depthwise_convolution_2d import depthwise_convolution_2d  # NOQA
 from chainer.functions.connection.dilated_convolution_2d import dilated_convolution_2d  # NOQA
 from chainer.functions.connection.embed_id import embed_id  # NOQA
 from chainer.functions.connection.linear import linear  # NOQA