--- a/chainer/links/__init__.py
+++ b/chainer/links/__init__.py
@@ -8,6 +8,7 @@ from chainer.links.connection import convolution_2d  # NOQA
 from chainer.links.connection import convolution_nd  # NOQA
 from chainer.links.connection import deconvolution_2d  # NOQA
 from chainer.links.connection import deconvolution_nd  # NOQA
+from chainer.links.connection import depthwise_convolution_2d  # NOQA
 from chainer.links.connection import dilated_convolution_2d  # NOQA
 from chainer.links.connection import embed_id  # NOQA
 from chainer.links.connection import gru  # NOQA
@@ -41,6 +42,7 @@ from chainer.links.connection.convolution_2d import Convolution2D  # NOQA
 from chainer.links.connection.convolution_nd import ConvolutionND  # NOQA
 from chainer.links.connection.deconvolution_2d import Deconvolution2D  # NOQA
 from chainer.links.connection.deconvolution_nd import DeconvolutionND  # NOQA
+from chainer.links.connection.depthwise_convolution_2d import DepthwiseConvolution2D  # NOQA
 from chainer.links.connection.dilated_convolution_2d import DilatedConvolution2D  # NOQA
 from chainer.links.connection.embed_id import EmbedID  # NOQA
 from chainer.links.connection.gru import GRU  # NOQA