--- a/chainer/links/connection/depthwise_convolution_2d.py
+++ b/chainer/links/connection/depthwise_convolution_2d.py
@@ -17,7 +17,8 @@ class DepthwiseConvolution2D(link.Link):
         in_channels (int): Number of channels of input arrays. If ``None``,
             parameter initialization will be deferred until the first forward
             data pass at which time the size will be determined.
-        channel_multiplier (int): Number of channel multiplier.
+        channel_multiplier (int): Channel multiplier number. Number of output
+            arrays equal ``in_channels * channel_multiplier``.
         ksize (int or pair of ints): Size of filters (a.k.a. kernels).
             ``ksize=k`` and ``ksize=(k, k)`` are equivalent.
         stride (int or pair of ints): Stride of filter applications.