--- a/chainer/links/connection/depthwise_convolution_2d.py
+++ b/chainer/links/connection/depthwise_convolution_2d.py
@@ -29,7 +29,7 @@ class DepthwiseConvolution2D(link.Link):
         bias (float): Initial bias value.
         nobias (bool): If ``True``, then this link does not use the bias term.
         initialW (4-D array): Initial weight value. If ``None``, then this
-            function uses to initialize ``wscale``.
+            function uses scaled Gaussian distribution to initialize weight.
             May also be a callable that takes ``numpy.ndarray`` or
             ``cupy.ndarray`` and edits its value.
         initial_bias (1-D array): Initial bias value. If ``None``, then this