--- a/docs/source/reference/links.rst
+++ b/docs/source/reference/links.rst
@@ -46,6 +46,10 @@ DeconvolutionND
 ~~~~~~~~~~~~~~~
 .. autoclass:: DeconvolutionND
 
+DepthwiseConvolutionND
+~~~~~~~~~~~~~~~~~~~~~~
+.. autoclass:: DepthwiseConvolutionND
+
 DilatedConvolution2D
 ~~~~~~~~~~~~~~~~~~~~
 .. autoclass:: DilatedConvolution2D