--- a/docs/source/reference/links.rst
+++ b/docs/source/reference/links.rst
@@ -46,9 +46,9 @@ DeconvolutionND
 ~~~~~~~~~~~~~~~
 .. autoclass:: DeconvolutionND
 
-DepthwiseConvolutionND
+DepthwiseConvolution2D
 ~~~~~~~~~~~~~~~~~~~~~~
-.. autoclass:: DepthwiseConvolutionND
+.. autoclass:: DepthwiseConvolution2D
 
 DilatedConvolution2D
 ~~~~~~~~~~~~~~~~~~~~