--- a/tests/chainer_tests/functions_tests/connection_tests/test_depthwise_convolution_2d.py
+++ b/tests/chainer_tests/functions_tests/connection_tests/test_depthwise_convolution_2d.py
@@ -91,11 +91,13 @@ class TestDepthwiseConvolution2DFunction(unittest.TestCase):
     def test_backward_cpu_nobias(self):
         self.check_backward(self.x, self.W, None, self.gy)
 
+    @attr.gpu
     @condition.retry(3)
     def test_backward_gpu(self):
         self.check_backward(cuda.to_gpu(self.x), cuda.to_gpu(self.W),
                             cuda.to_gpu(self.b), cuda.to_gpu(self.gy))
 
+    @attr.gpu
     @condition.retry(3)
     def test_backward_gpu_nobias(self):
         self.check_backward(cuda.to_gpu(self.x), cuda.to_gpu(self.W),