… tests/chainer_tests/links_tests/connection_tests/test_depthwise_convolution_2d.py