This is PR for supporting depthwise convolution 2D.
Main usage for this layer is separable convolution.

(http://www.di.ens.fr/data/publications/papers/phd_sifre.pdf)
(https://arxiv.org/abs/1610.02357)

After merge this PR I wil restart separable convolution #1956.