@fukatani 2017-01-23
Thanks for your advice. :smiley: 
I have struggled to implement this PR without for-loop.

Now, I completed implementation.
But the part of GPU test need `cupy.matmul` for float16 and float64. (#2151)