@fukatani 2017-03-05
I confirmed GPU test.
Could you check this PR?