@fukatani 2017-03-06
Thank you for your comments. I fixed them.