@fukatani 2017-03-16
Thanks for your attentive review.
It will be helpful if you can see the identity with the paper.