@fukatani 2017-03-17
Is this sentence natural English?
```
If ``None``, then this function uses to initialize ``bias``.
```
I think the following phrase is more correct.
```
If ``None``, then this link uses ``bias`` to initialize bias.
```

If so, I fix depthwise doc.
And some doc of links use same phrase, I'll fix them.