@fukatani 2017-03-22
I measured forward computation performance with the following script.

with transpose: 0.00181772509765625 sec
without transpose: 0.00179154296875 sec

The impact on backward computation time is even less.

```python
import cupy
from chainer.utils import conv

sy, sx = 1, 1
ph, pw = 0, 0

x = cupy.arange(300000).reshape(10, 3, 100, 100)
col = conv.im2col_gpu(x, 100, 100, sy, sx, ph, pw)
W1 = cupy.arange(90000).reshape(3, 3, 100, 100)
W2 = cupy.arange(90000).reshape(3, 100, 100, 3)

out_c, _, kh, kw = W1.shape
n, c, h, w = x.shape

out_h = conv.get_conv_outsize(h, kh, sy, ph)
out_w = conv.get_conv_outsize(w, kw, sx, pw)


def with_transpose():
    col = conv.im2col_gpu(x, kh, kw, sy, sx, ph, pw)
    B, C, KY, KX, IY, IX = col.shape
    D = W1.shape[0]  # (D, C, KY, KX)
    c_ = col.transpose(1, 0, 4, 5, 2, 3).reshape((C, B * IY * IX, KY * KX))
    w_ = W1.transpose(1, 2, 3, 0).reshape((C, KY * KX, D))

    y = cupy.matmul(c_, w_).astype(x.dtype, copy=False)
    return y.reshape((C, B, IY * IX, D)).transpose(1, 0, 3, 2).reshape((B, C * D, IY, IX))


def without_transpose():
    col = conv.im2col_gpu(x, kh, kw, sy, sx, ph, pw)
    B, C, KY, KX, IY, IX = col.shape
    D = W2.shape[3]  # (D, C, KY, KX)
    c_ = col.transpose(1, 0, 4, 5, 2, 3).reshape((C, B * IY * IX, KY * KX))
    w_ = W2.reshape((C, KY * KX, D))

    y = cupy.matmul(c_, w_).astype(x.dtype, copy=False)
    return y.reshape((C, B, IY * IX, D)).transpose(1, 0, 3, 2).reshape((B, C * D, IY, IX))


with_transpose()  # avoid initialize
without_transpose()  # avoid initialize

cnt = 1000
stream = cupy.cuda.Stream.null
start = stream.record()
for i in range(cnt):
    with_transpose()
end = stream.record()
end.synchronize()

print(cupy.cuda.get_elapsed_time(start, end) / 1000 / cnt)

start = stream.record()
for i in range(cnt):
    without_transpose()
end = stream.record()
end.synchronize()
print(cupy.cuda.get_elapsed_time(start, end) / 1000 / cnt
```