@delta2323 2017-03-23
Thank you for your experiment! As I confirm that the introduction of transpose does not affect too much, I agree with your original implementation.