@delta2323 2017-03-23
LGTM, I am running GPU tests. 

@mattya Should I merge this PR if I confirm this PR is merge-ready?