@delta2323 2017-03-23
As you can see, currently Travis tests of this PR fail. But this error seems independent of this PR as I encountered the same problem in another PR #1754. I am now investigating the problem.