@mattya 2017-03-30
@delta2323 I think it's ok to merge after solving the Travis problem.