@delta2323 2017-04-03
As you can see, there occurs the following errors in Travis. This is because old NumPy (< 1.10) does not have `matmul` (see [reference](https://docs.scipy.org/doc/numpy/reference/generated/numpy.matmul.html)). As Chainer support NumPy older than 1.10, we should make this function work with it. We have several choices:

* Implement this function without `numpy.matmul`
* Check the version of NumPy and switch the implementation depending on the version.
* Check the version of NumPy and raise an error if NumPy is older than 1.10.

Could you consider the implement this function without `matmul`? If it is difficult, I think the workaround is to take the third option and support older NumPy in another PR.

```
======================================================================
ERROR: test_forward_cpu (test_simplified_dropconnect.TestSimplifiedDropconnect_param_0)  parameter: {'train': True, 'ratio': 0.0, 'W_dtype': <type 'numpy.float16'>, 'x_dtype': <type 'numpy.float16'>}
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/travis/build/pfnet/chainer/tests/chainer_tests/functions_tests/noise_tests/test_simplified_dropconnect.py", line 56, in test_forward_cpu
    self.check_forward(self.x, self.W, self.b)
  File "/home/travis/build/pfnet/chainer/tests/chainer_tests/functions_tests/noise_tests/test_simplified_dropconnect.py", line 52, in check_forward
    self.train)
  File "/home/travis/virtualenv/python2.7.9/lib/python2.7/site-packages/chainer/functions/noise/simplified_dropconnect.py", line 148, in simplified_dropconnect
    return SimplifiedDropconnect(ratio, mask)(x, W, b)
  File "/home/travis/virtualenv/python2.7.9/lib/python2.7/site-packages/chainer/function.py", line 199, in __call__
    outputs = self.forward(in_data)
  File "/home/travis/virtualenv/python2.7.9/lib/python2.7/site-packages/chainer/functions/noise/simplified_dropconnect.py", line 61, in forward
    y = xp.matmul(W, x[:, :, None])
AttributeError: 'module' object has no attribute 'matmul'
```