@fukatani 2017-04-03
`einsum` can be a substitute for `matmul`.

Performance is almost the same.
As long as I benchmarked, `matmul` is only 0.5% faster (in numpy.1.11).