@delta2323 2017-04-05
Thank you very much! I will run GPU tests and will merge this PR if all tests pass.