@delta2323 2017-04-06
GPU test failed because of unit tests that seem irrelevant to this PR. Could you try to rebase or merge the latest master branch?